add llmPanel
